Refactor URL aware authenticators.
